+++ StabilityMatrix.Avalonia/ViewModels/PackageManager/PackageExtensionBrowserViewModel.cs
@@ -690,6 +690,76 @@ public partial class PackageExtensionBrowserViewModel : ViewModelBase, IDisposab
         }
     }
 
+    [RelayCommand]
+    private async Task InstallExtensionManualAsync()
+    {
+        var textField = new TextBoxField
+        {
+            Label = "Extension URL",
+            Validator = text =>
+            {
+                if (string.IsNullOrWhiteSpace(text))
+                    throw new DataValidationException("URL is required");
+
+                if (!Uri.TryCreate(text, UriKind.Absolute, out _))
+                    throw new DataValidationException("Invalid URL format");
+            },
+        };
+        var dialog = DialogHelper.CreateTextEntryDialog("Manual Extension Install", "", [textField]);
+
+        if (await dialog.ShowAsync() != ContentDialogResult.Primary)
+            return;
+
+        var url = textField.Text.Trim();
+        if (string.IsNullOrWhiteSpace(url))
+            return;
+
+        // check if have enough parts
+        if (url.Split('/').Length < 5)
+        {
+            notificationService.Show("Invalid URL", "The provided URL does not contain enough information.");
+            return;
+        }
+
+        // get the author from github url
+        var author = url.Split('/')[3];
+        // get the title from the url
+        var title = url.Split('/')[4].Replace(".git", "");
+        // create a new PackageExtension
+        var packageExtension = new PackageExtension
+        {
+            Author = author,
+            Title = title,
+            Reference = new Uri(url),
+            IsInstalled = false,
+            InstallType = "git-clone",
+            Files = [new Uri(url)],
+        };
+
+        var steps = new List<IPackageStep>
+        {
+            new InstallExtensionStep(
+                PackagePair!.BasePackage.ExtensionManager!,
+                PackagePair.InstalledPackage,
+                packageExtension
+            ),
+        };
+
+        var runner = new PackageModificationRunner
+        {
+            ShowDialogOnStart = true,
+            ModificationCompleteTitle = "Installed Extensions",
+            ModificationCompleteMessage = "Finished installing extensions",
+        };
+        EventManager.Instance.OnPackageInstallProgressAdded(runner);
+
+        await runner.ExecuteSteps(steps);
+
+        ClearSelection();
+
+        RefreshBackground();
+    }
+
     public void RefreshBackground()
     {
         RefreshCore()