+++ CHANGELOG.md
@@ -6,8 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2.0.0.html).
 
 ## v2.15.0-pre.2
+### Added
+- Added Manual Install button for installing Package extensions that aren't in the indexes
+- Added Next and Previous buttons to the Civitai details page to navigate between results
+### Changed
+- Brought back the "size remaining after download" tooltip in the new Civitai details page
 ### Fixed
 - Fixed Inference custom step (e.g. HiresFix) Samplers potentially sharing state with other card UIs like model browser.
+- Fixed extension manager failing to install extensions due to incorrect clone directory
 
 ## v2.15.0-pre.1
 ### Added